[gl] disable glutin and surfman
